requirements.txt
uvicorn==0.22.0
fastapi==0.95.1
requests==2.28.1
psycopg2-binary==2.9.9